Introduce SmallVector implementation.
Rationale
---------

Even if LargInt<> are Plain Old Datatypes, Integer is not a POD because of the variant.
Consequently, Node_t and Edge_t are not PODs and require initialization by an
(implicit) constructor.

This means that all unused GraphVector slots are initialized with Node_t<Integer>
containing LargeInt<1>. Then, in GraphTemplate's methods, Node_t.set() sets the
variant to LargeInt<span> for the span currently used.

Replacement of GraphVector
--------------------------

The new implementation still use a fixed size allocation on the stack, but
leverage std::aligned_storage to make a buffer with no initialization cost.
Slot are filled by a placement new through the emplace_back interface.